test(select): migrate basic
